SportacuzApi/models/token.py
import jwt
import datetime

from Sportacuz.config import key


def encode_auth_token(user_id):
    """
    Encode token for user
    :param user_id:
    :return: string
    """
    try:
        payload = {
            'exp': datetime.datetime.utcnow() + datetime.timedelta(days=1),
            'iat': datetime.datetime.utcnow(),
            'sub': user_id
        }
        return jwt.encode(
            payload,
            key,
            algorithm='HS256'
        )
    except Exception as e:
        return e


def decode_auth_token(auth_token):
    """
    Decode the Auth token
    :param auth_token:
    :return: integer|string
    """
    try:
        payload = jwt.decode(auth_token, key)
        return payload['sub']
    except jwt.ExpiredSignatureError:
        return 'Signature expired. Please login again.'
    except jwt.InvalidTokenError:
        return 'Invalid token. Please log in again.'
