requirements.txt
alembic==1.0.0
aniso8601==3.0.2
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
flake8==3.5.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.2.1
flask-restplus==0.11.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
pipenv==2018.7.1
psycopg2==2.7.5
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.6.4
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.10
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
